[Tweak] Create own component for registerForm - TKDATA-155
