+++ BilparkApp/src/components/UserRegister/UserRegisterScreen.js
@@ -1,77 +1,14 @@
 import React from 'react';
-import t from 'tcomb-form-native';
 import { connect } from 'react-redux';
-import { Text, Modal, StyleSheet, View, Button, ScrollView, Dimensions } from 'react-native';
+import { StyleSheet, ScrollView } from 'react-native';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view';
-import { registerFetchData, registerModalVisible, registerResetOptionUpdateValue, registerUpdateValue } from '../../actions/registerUser';
+import { registerModalVisible } from '../../actions/registerUser';
 import UserRegisterModal from './UserRegisterModal';
-
-
-const width = Dimensions.get('window').width;
-
-t.form.Form.stylesheet.textbox.normal.color = '#000';
-t.form.Form.stylesheet.formGroup.normal.width = (width - 20);
-t.form.Form.stylesheet.formGroup.error.width = (width - 20);
-t.form.Form.stylesheet.textbox.normal.height = 40;
-t.form.Form.stylesheet.textbox.error.height = 40;
-t.form.Form.stylesheet.textbox.normal.fontSize = 20;
-t.form.Form.stylesheet.textbox.error.fontSize = 20;
-t.form.Form.stylesheet.textbox.normal.backgroundColor = '#fff';
-t.form.Form.stylesheet.textbox.error.backgroundColor = '#fff';
-t.form.Form.stylesheet.textbox.error.borderWidth = 3;
-t.form.Form.stylesheet.errorBlock.color = '#db2b1e';
-t.form.Form.stylesheet.controlLabel.normal.color = '#fff';
-t.form.Form.stylesheet.controlLabel.error.color = '#fff';
-
-const mailCheck = t.refinement(t.String, (email) => {
-  const reg = /[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/;
-  return reg.test(email);
-});
-
-const passwordCheck = t.refinement(t.String, (pass) => {
-  return (pass.length >= 8);
-});
-
-const stringCheck = t.refinement(t.String, (string) => {
-  const reg = /[a-zA-Z0-9-]/;
-  return reg.test(string);
-});
-
-const User = t.struct({
-  Email: mailCheck,
-  Fname: stringCheck,
-  Lname: stringCheck,
-  Address: stringCheck,
-  Password: passwordCheck,
-});
-
-const Form = t.form.Form;
+import UserRegisterForm from './UserRegisterForm';
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#002776',
-    padding: 10,
-  },
-  welcome: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10,
-  },
   keyboard: {
     backgroundColor: '#002776',
-  },
-  modal: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 10,
-    backgroundColor: '#002776',
-  },
-  modalText: {
-    color: '#fff',
   },
 });
 
@@ -81,20 +18,6 @@ class registerScreen extends React.Component {
     this.props.visibleModal(false);
   }
 
-  onChange(values) {
-    this.props.addValues(values);
-  }
-
-  async handleSubmit() {
-    const value = this.form.getValue();
-    // use that ref to get the form value
-    this.props.resetOptions();
-    if (value) {
-      this.props.fetchData(value);
-    }
-  }
-
-
   render() {
     return (
       <KeyboardAwareScrollView
@@ -103,21 +26,7 @@ class registerScreen extends React.Component {
         contentContainerStyle={styles.keyboard}
       >
         <ScrollView>
-          <View style={styles.container}>
-            <Form
-              ref={c => this.form = c}
-              type={User}
-              options={this.props.options}
-              value={this.props.values}
-              onChange={value => this.onChange(value)}
-            />
-            <Button
-              title="Registrer"
-              onPress={() => {
-                this.handleSubmit();
-              }}
-            />
-          </View>
+          <UserRegisterForm />
         </ScrollView>
         <UserRegisterModal />
       </KeyboardAwareScrollView>
@@ -125,19 +34,14 @@ class registerScreen extends React.Component {
   }
 }
 
-const mapStateToProps = (state) => {
+const mapStateToProps = () => {
   return {
-    options: state.options,
-    values: state.values,
   };
 };
 
 const mapDispatchToProps = (dispatch) => {
   return {
-    fetchData: values => dispatch(registerFetchData(values)),
     visibleModal: bool => dispatch(registerModalVisible(bool)),
-    resetOptions: () => dispatch(registerResetOptionUpdateValue()),
-    addValues: value => dispatch(registerUpdateValue(value)),
   };
 };
